requirements.txt
mysql-connector-python>=8.0.0
google-generativeai>=0.3.0
tabulate>=0.9.0
requests>=2.31.0
groq>=0.4.0
streamlit>=1.30.0
pandas>=2.0.0
